requirements.txt
aiokafka==0.12.0
asyncpg==0.30.0
kafka_python==2.0.4
python-dotenv==1.0.1
Requests==2.32.3
